+++ src/Codeception/Lib/Driver/Db.php
@@ -58,8 +58,6 @@ class Db
                 return new PostgreSql($dsn, $user, $password);
             case 'mssql':
                 return new MsSql($dsn, $user, $password);
-            case 'oracle':
-                return new Oracle($dsn, $user, $password);
             case 'sqlsrv':
                 return new SqlSrv($dsn, $user, $password);
             case 'oci':
