+++ src/Codeception/Lib/Driver/Oci.php
@@ -1,8 +1,37 @@
 <?php
 namespace Codeception\Lib\Driver;
 
-class Oci extends Oracle
+class Oci extends Db
 {
+
+    public function cleanup()
+    {
+        $this->dbh->exec(
+            "BEGIN
+                            FOR i IN (SELECT trigger_name FROM user_triggers)
+                              LOOP
+                                EXECUTE IMMEDIATE('DROP TRIGGER ' || user || '.' || i.trigger_name);
+                              END LOOP;
+                          END;"
+        );
+        $this->dbh->exec(
+            "BEGIN
+                            FOR i IN (SELECT table_name FROM user_tables)
+                              LOOP
+                                EXECUTE IMMEDIATE('DROP TABLE ' || user || '.' || i.table_name || ' CASCADE CONSTRAINTS');
+                              END LOOP;
+                          END;"
+        );
+        $this->dbh->exec(
+            "BEGIN
+                            FOR i IN (SELECT sequence_name FROM user_sequences)
+                              LOOP
+                                EXECUTE IMMEDIATE('DROP SEQUENCE ' || user || '.' || i.sequence_name);
+                              END LOOP;
+                          END;"
+        );
+    }
+
     /**
      * SQL commands should ends with `//` in the dump file
      * IF you want to load triggers too.
